Take serial port comms speed up a notch.  Leave some options
While trying this codebase on several new clock modules, the 9k6bps option
was found not work work!

It seems that there is an extra software step needed to perform OSCCAL
calibration on a atmega8 device running at something other than 1Mhz.
So this may be the reason.

Strangely enough, slowing the speed down did not help anything, so I have
gone up to the next usable option